# aulas/aula10/swagger.yaml
openapi: 3.0.0
info:
  title: API Loja Virtual
  version: 0.0.1
tags:
  - produtos
paths:
  /produtos:
    post:
      tags:
        - produtos
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Produto'
      responses:
        201:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'
        422:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'
    get:
      security:
        - JWTAuth: []
      tags:
        - produtos
      responses:
        200:
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Produto'     
  /produtos/{id}:
    get:
      tags:
        - produtos
      parameters:
        - in: path
          name: id
          type: string  
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'  
        400:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'        
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'        
    put:
      tags:
        - produtos
      parameters:
        - in: path
          name: id
          type: string  
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Produto'
      responses:
        200:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Produto'  
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'        
        422:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'       
    delete:
      tags:
        - produtos
      parameters:
        - in: path
          name: id
          type: string  
      responses:
        204:
          content:
        400:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'        
        404:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Mensagem'                                                               
components:
  securitySchemes:
    JWTAuth:
      type: apiKey
      in: header
      name: Authorization
  schemas:
    Produto:
      type: object
      properties:
        nome:
          type: string
        preco:
          type: number
    Mensagem:
      type: object
      properties:
        msg:
          type: string